Updates Algorand SDK reference to Algorand .NET SDK 2 by FrankSzendzielarz

+++ src/Yieldly/V1/AsaStakingPool.cs
@@ -49,7 +49,7 @@ namespace Yieldly.V1 {
 		/// </summary>
 		public async Task RefreshAsync() {
 
-			Application = await Client.DefaultApi.ApplicationsAsync((int)ApplicationId);
+			Application = await Client.DefaultApi.ApplicationsAsync(ApplicationId);
 		}
 
 		/// <summary>
